requirements.txt
pylint
django-debug-toolbar
django-extensions
pytest
pytest-sugar
pytest-django
pytest-factoryboy
pytest-lazy-fixture
coverage
django-coverage-plugin
pytest-cov
django-cleanup
django-dbtemplates
tox
sphinx
sphinx-autobuild
sphinx-rtd-theme
furo
sphinxcontrib-images